Get rid of FlinkRangeMap and make MapeState main implementation
