+++ engine/flink/components/base-unbounded/src/main/scala/pl/touk/nussknacker/engine/flink/util/transformer/aggregate/AggregatorFunction.scala
@@ -1,21 +1,24 @@
 package pl.touk.nussknacker.engine.flink.util.transformer.aggregate
 
-import org.apache.flink.api.common.functions.RuntimeContext
-import org.apache.flink.api.common.state.ValueStateDescriptor
+import cats.data.NonEmptyList
+import org.apache.flink.api.common.functions.{OpenContext, RichFunction, RuntimeContext}
+import org.apache.flink.api.common.state.{MapState, MapStateDescriptor, ValueState, ValueStateDescriptor}
 import org.apache.flink.api.common.typeinfo.TypeInformation
+import org.apache.flink.api.java.typeutils.ListTypeInfo
 import org.apache.flink.streaming.api.TimerService
 import org.apache.flink.streaming.api.functions.KeyedProcessFunction
 import org.apache.flink.util.Collector
 import pl.touk.nussknacker.engine.api.{Context => NkContext, NodeId, ValueWithContext}
 import pl.touk.nussknacker.engine.api.runtimecontext.EngineRuntimeContext
 import pl.touk.nussknacker.engine.api.typed.typing.TypingResult
-import pl.touk.nussknacker.engine.flink.api.state.{LatelyEvictableStateFunction, StateHolder}
 import pl.touk.nussknacker.engine.flink.util.keyed.KeyEnricher
-import pl.touk.nussknacker.engine.flink.util.orderedmap.FlinkRangeMap
-import pl.touk.nussknacker.engine.flink.util.orderedmap.FlinkRangeMap._
+import pl.touk.nussknacker.engine.util
 import pl.touk.nussknacker.engine.util.KeyedValue
+import pl.touk.nussknacker.engine.util.metrics.{MetricIdentifier, MetricsProviderForScenario}
+import pl.touk.nussknacker.engine.util.metrics.common.naming.nodeIdTag
 
-import scala.language.higherKinds
+import java.util.Comparator
+import scala.jdk.CollectionConverters._
 
 // This is the real SlidingWindow with slide = 1min - moving with time for each key. It reduce on each emit and store
 // partial aggregations for each minute, instead of storing aggregates for each slide. This implementation doesn't emit event
@@ -23,24 +26,23 @@ import scala.language.higherKinds
 // Flinks SlidingWindows are something that is often called HoppingWindow. They consume a lot of memory because each element
 // is stored in multiple windows with different offsets.
 // NOTE: it would be much cleaner if we evaluated aggregateBy here. However, FLINK-10250 prevents us from doing this and we *have* to compute it before
-class AggregatorFunction[MapT[K, V]](
+class AggregatorFunction(
     protected val aggregator: Aggregator,
     protected val timeWindowLengthMillis: Long,
     override val nodeId: NodeId,
     protected val aggregateElementType: TypingResult,
-    protected override val aggregateTypeInformation: TypeInformation[AnyRef],
+    protected val aggregateTypeInformation: TypeInformation[AnyRef],
     val convertToEngineRuntimeContext: RuntimeContext => EngineRuntimeContext
-)(implicit override val rangeMap: FlinkRangeMap[MapT])
-    extends LatelyEvictableStateFunction[
-      ValueWithContext[KeyedValue[AnyRef, AnyRef]],
-      ValueWithContext[AnyRef],
-      MapT[Long, AnyRef],
-      AnyRef
-    ]
-    with AggregatorFunctionMixin[MapT] {
+) extends KeyedProcessFunction[AnyRef, ValueWithContext[KeyedValue[AnyRef, AnyRef]], ValueWithContext[AnyRef]]
+    with AggregatorFunctionMixin {
 
   type FlinkCtx =
     KeyedProcessFunction[AnyRef, ValueWithContext[KeyedValue[AnyRef, AnyRef]], ValueWithContext[AnyRef]]#Context
+
+  override def open(openContext: OpenContext): Unit = {
+    super.open(openContext)
+    initState()
+  }
 
   override def processElement(
       value: ValueWithContext[KeyedValue[AnyRef, AnyRef]],
@@ -50,11 +52,99 @@ class AggregatorFunction[MapT[K, V]](
     handleNewElementAdded(value, ctx.timestamp(), ctx.timerService(), out)
   }
 
+  override def onTimer(
+      timestamp: Long,
+      ctx: KeyedProcessFunction[
+        AnyRef,
+        ValueWithContext[KeyedValue[AnyRef, AnyRef]],
+        ValueWithContext[AnyRef]
+      ]#OnTimerContext,
+      out: Collector[ValueWithContext[AnyRef]]
+  ): Unit = {
+    handleOnTimer(timestamp, ctx.timerService())
+  }
+
 }
 
-trait AggregatorFunctionMixin[MapT[K, V]] extends AggregatorFunctionBase { self: StateHolder[MapT[Long, AnyRef]] =>
+trait AggregatorFunctionMixin extends RichFunction {
 
-  protected implicit def rangeMap: FlinkRangeMap[MapT]
+  protected def convertToEngineRuntimeContext: RuntimeContext => EngineRuntimeContext
+
+  protected lazy val engineRuntimeContext: EngineRuntimeContext = convertToEngineRuntimeContext(getRuntimeContext)
+
+  def nodeId: NodeId
+
+  protected def name: String = "aggregator"
+
+  protected def tags: Map[String, String] = Map(nodeIdTag -> nodeId.id)
+
+  protected lazy val metricsProvider: MetricsProviderForScenario = engineRuntimeContext.metricsProvider
+
+  protected lazy val timeHistogram: util.metrics.Histogram =
+    metricsProvider.histogram(MetricIdentifier(NonEmptyList.of(name, "time"), tags), 10)
+
+  // this metric does *not* calculate histogram of sizes of maps in the whole state,
+  // but of those that are processed, so "hot" keys would be counted much more often.
+  protected lazy val retrievedBucketsHistogram: util.metrics.Histogram =
+    metricsProvider.histogram(MetricIdentifier(NonEmptyList.of(name, "retrievedBuckets"), tags), 10)
+
+  protected def minimalResolutionMs: Long = 60000L
+
+  protected def allowedOutOfOrderMs: Long = timeWindowLengthMillis
+
+  protected val aggregator: Aggregator
+
+  protected def timeWindowLengthMillis: Long
+
+  protected def aggregateElementType: TypingResult
+
+  protected val outputType: TypingResult = aggregator
+    .computeOutputType(aggregateElementType)
+    .valueOr(e => throw new IllegalArgumentException("Failed to compute output type: " + e))
+
+  protected def aggregateTypeInformation: TypeInformation[AnyRef]
+
+  @transient
+  protected var bucketsState: BucketsState = _
+
+  @transient
+  protected var latestEvictionTimeForKey: ValueState[java.lang.Long] = _
+
+  protected def handleNewElementAdded(
+      value: ValueWithContext[KeyedValue[AnyRef, AnyRef]],
+      timestamp: Long,
+      timeService: TimerService,
+      out: Collector[ValueWithContext[AnyRef]]
+  ): Unit = {
+    val start = System.nanoTime()
+    addElementToState(value, timestamp, timeService, out)
+    val finalVal = computeFinalValue(timestamp, bucketsState.keys)
+    timeHistogram.update(System.nanoTime() - start)
+    out.collect(ValueWithContext(finalVal, KeyEnricher.enrichWithKey(value.context, value.value)))
+
+  }
+
+  protected def addElementToState(
+      value: ValueWithContext[KeyedValue[AnyRef, AnyRef]],
+      timestamp: Long,
+      timeService: TimerService,
+      out: Collector[ValueWithContext[AnyRef]]
+  ): Unit = {
+    val bucketTs   = computeTimestampToStore(timestamp)
+    val newElement = value.value.value.asInstanceOf[aggregator.Element]
+
+    val currentAggregate = Option(bucketsState.get(bucketTs))
+      .getOrElse(aggregator.createAccumulator())
+      .asInstanceOf[aggregator.Aggregate]
+
+    if (!aggregator.isNeutralForAccumulator(newElement, currentAggregate)) {
+      val newAggregate = aggregator.add(newElement, currentAggregate)
+      bucketsState.put(bucketTs, newAggregate)
+    }
+
+    doMoveEvictionTime(bucketTs + timeWindowLengthMillis, timeService)
+    handleElementAddedToState(bucketTs, newElement, value.context, timeService, out)
+  }
 
   // for extending classes purpose
   protected def handleElementAddedToState(
@@ -65,104 +155,131 @@ trait AggregatorFunctionMixin[MapT[K, V]] extends AggregatorFunctionBase { self:
       out: Collector[ValueWithContext[AnyRef]]
   ): Unit = {}
 
-  protected def handleNewElementAdded(
-      value: ValueWithContext[KeyedValue[AnyRef, AnyRef]],
-      timestamp: Long,
-      timeService: TimerService,
-      out: Collector[ValueWithContext[AnyRef]]
-  ): Unit = {
-    val start    = System.nanoTime()
-    val newState = addElementToState(value, timestamp, timeService, out)
-    val finalVal = computeFinalValue(newState, timestamp)
-    timeHistogram.update(System.nanoTime() - start)
-    out.collect(ValueWithContext(finalVal, KeyEnricher.enrichWithKey(value.context, value.value)))
+  protected def computeFinalValue(timestamp: Long, bucketKeys: java.util.List[java.lang.Long]): AnyRef = {
+    val rangeStart = timestamp - timeWindowLengthMillis + 1
+    var count      = 0
+
+    val foldedState = bucketKeys.asScala
+      .filter(ts => ts >= rangeStart && ts <= timestamp)
+      .foldLeft(aggregator.createAccumulator()) { (acc, key) =>
+        count += 1
+        aggregator.merge(acc, bucketsState.get(key))
+      }
+
+    retrievedBucketsHistogram.update(count)
+    aggregator.alignToExpectedType(aggregator.getResult(foldedState), outputType)
   }
 
-  protected def addElementToState[T <: AnyRef](
-      value: ValueWithContext[KeyedValue[T, AnyRef]],
-      timestamp: Long,
-      timeService: TimerService,
-      out: Collector[ValueWithContext[AnyRef]]
-  ): MapT[Long, aggregator.Aggregate] = {
-    val newElementInStateTimestamp  = computeTimestampToStore(timestamp)
-    val newElement                  = value.value.value.asInstanceOf[aggregator.Element]
-    val (newState, stateWasChanged) = computeNewState(newElementInStateTimestamp, newElement)
+  // Timer management — duplicated from LatelyEvictableStateFunctionMixin (~20 lines, stable code).
+  // Not extracted to shared trait to avoid touching LatelyEvictableStateFunctionMixin which is used
+  // by unrelated components (UnionWithMemoTransformer, TransformStateTransformer).
+  protected def doMoveEvictionTime(time: Long, timeService: TimerService): Unit = {
+    val latestEvictionTimeValue = latestEvictionTimeForKey.value()
+    val maxEvictionTime = if (latestEvictionTimeValue == null || time > latestEvictionTimeValue) {
+      time
+    } else {
+      latestEvictionTimeValue.longValue()
+    }
 
-    if (stateWasChanged) {
-      updateState(
-        newState.asInstanceOf[MapT[Long, AnyRef]],
-        newElementInStateTimestamp + timeWindowLengthMillis,
-        timeService
+    if (latestEvictionTimeValue == null) {
+      timeService.registerEventTimeTimer(maxEvictionTime)
+    }
+
+    latestEvictionTimeForKey.update(maxEvictionTime)
+  }
+
+  protected def handleOnTimer(timestamp: Long, timerService: TimerService): Unit = {
+    val latestEvictionTimeValue = latestEvictionTimeForKey.value()
+    val noLaterEventsArrived    = latestEvictionTimeValue == timestamp
+
+    if (noLaterEventsArrived) {
+      bucketsState.clear()
+      latestEvictionTimeForKey.update(null)
+    } else if (latestEvictionTimeValue != null) {
+      evictOldBuckets(timestamp)
+      timerService.registerEventTimeTimer(latestEvictionTimeValue)
+    }
+  }
+
+  private def evictOldBuckets(currentWatermark: Long): Unit = {
+    val cutoff = currentWatermark - timeWindowLengthMillis + 1 - allowedOutOfOrderMs
+    bucketsState.removeOlderThan(cutoff)
+  }
+
+  private def computeTimestampToStore(timestamp: Long): Long = {
+    (timestamp / minimalResolutionMs) * minimalResolutionMs
+  }
+
+  protected def initState(): Unit = {
+    val mapState = getRuntimeContext.getMapState(
+      new MapStateDescriptor[java.lang.Long, AnyRef](
+        "buckets",
+        TypeInformation.of(classOf[java.lang.Long]),
+        aggregateTypeInformation
       )
-    } else {
-      doMoveEvictionTime(newElementInStateTimestamp + timeWindowLengthMillis, timeService)
-    }
-    handleElementAddedToState(newElementInStateTimestamp, newElement, value.context, timeService, out)
-    retrievedBucketsHistogram.update(newState.toScalaMapRO.size)
-    newState
-  }
-
-  protected def computeFinalValue(newState: MapT[Long, aggregator.Aggregate], timestamp: Long): AnyRef = {
-    aggregator.alignToExpectedType(computeFoldedAggregatedValue(newState, timestamp), outputType)
-  }
-
-  protected def computeFoldedAggregatedValue(state: MapT[Long, aggregator.Aggregate], timestamp: Long): AnyRef = {
-    val newStateTruncatedToTimestamp = stateForTimestampToRead(state, timestamp)
-    val foldedState = if (newStateTruncatedToTimestamp.toScalaMapRO.isEmpty) {
-      aggregator.createAccumulator()
-    } else {
-      // here we need ordering for start, end etc.
-      newStateTruncatedToTimestamp.toScalaMapRO.values.reduce(aggregator.merge)
-    }
-    aggregator.getResult(foldedState)
-  }
-
-  // Return state after adding newValue and information about if state was changed or not
-  private def computeNewState(
-      newElementInStateTimestamp: Long,
-      newValue: aggregator.Element
-  ): (MapT[Long, aggregator.Aggregate], Boolean) = {
-    val current: MapT[Long, aggregator.Aggregate] =
-      stateForTimestampToSave(readStateOrInitial(), newElementInStateTimestamp)
-    val currentAggregate = current.toScalaMapRO
-      .getOrElse(newElementInStateTimestamp, aggregator.createAccumulator())
-      .asInstanceOf[aggregator.Aggregate]
-
-    // We do not create aggregate and add to it neutral element to avoid unnecessary buckets in our state
-    if (aggregator.isNeutralForAccumulator(newValue, currentAggregate)) {
-      (current, false)
-    } else {
-      val newAggregate = aggregator.add(newValue, currentAggregate).asInstanceOf[aggregator.Aggregate]
-      (current.updated(newElementInStateTimestamp, newAggregate), true)
-    }
-  }
-
-  protected def stateForTimestampToSave[T](stateValue: MapT[Long, T], timestamp: Long): MapT[Long, T] = {
-    stateValue.from(
-      timestamp - timeWindowLengthMillis + 1 - allowedOutOfOrderMs
-    ) // we must have exclusive range like Flink's sliding/tumbling have
-  }
-
-  protected def stateForTimestampToRead[T](stateValue: MapT[Long, T], timestamp: Long): MapT[Long, T] = {
-    stateForTimestampToReadUntilEnd(stateValue, timestamp).toRO(
-      timestamp
-    ) // we must have exclusive range like Flink's sliding/tumbling have
-  }
-
-  protected def stateForTimestampToReadUntilEnd[T](stateValue: MapT[Long, T], timestamp: Long): MapT[Long, T] = {
-    stateValue.fromRO(timestamp - timeWindowLengthMillis + 1)
-  }
-
-  protected def readStateOrInitial(): MapT[Long, aggregator.Aggregate] =
-    Option(readState().asInstanceOf[MapT[Long, aggregator.Aggregate]])
-      .getOrElse(rangeMap.empty[Long, aggregator.Aggregate])
-
-  override protected def stateDescriptor: ValueStateDescriptor[MapT[Long, AnyRef]] =
-    new ValueStateDescriptor[MapT[Long, AnyRef]](
-      "state",
-      rangeMap.typeInformation[Long, AnyRef](TypeInformation.of(classOf[Long]), aggregateTypeInformation)
+    )
+    val keysState = getRuntimeContext.getState(
+      new ValueStateDescriptor[java.util.List[java.lang.Long]](
+        "keys",
+        new ListTypeInfo(TypeInformation.of(classOf[java.lang.Long]))
+      )
     )
 
-  protected def aggregateTypeInformation: TypeInformation[AnyRef]
+    bucketsState = new BucketsState(mapState, keysState)
+    initEvictionTimeState()
+  }
+
+  protected def initEvictionTimeState(): Unit = {
+    latestEvictionTimeForKey = getRuntimeContext.getState[java.lang.Long](
+      new ValueStateDescriptor[java.lang.Long]("timers", classOf[java.lang.Long])
+    )
+  }
+
+  /**
+   * Manages bucket data for MapState-based sliding window aggregators.
+   *
+   * Maintains a separate ValueState with bucket timestamps (keys) to avoid
+   * expensive prefix scans over RocksDB MapState for key-only operations
+   * like range checks, max computation, and eviction decisions.
+   */
+  class BucketsState(
+      private val mapState: MapState[java.lang.Long, AnyRef],
+      private val keysState: ValueState[java.util.List[java.lang.Long]]
+  ) {
+
+    // Keys are always ordered. We order them before saving them to the state
+    def keys: java.util.List[java.lang.Long] = keysState.value() match {
+      case null => new java.util.ArrayList[java.lang.Long]()
+      case v    => v
+    }
+
+    def get(key: java.lang.Long): AnyRef =
+      mapState.get(key)
+
+    def put(key: java.lang.Long, value: AnyRef): Unit = {
+      mapState.put(key, value)
+      val currentKeys = keys
+      if (!currentKeys.contains(key)) {
+        currentKeys.add(key)
+        currentKeys.sort(Comparator.naturalOrder())
+        // We keep sorted list of keys
+        keysState.update(currentKeys)
+      }
+    }
+
+    def removeOlderThan(cutoff: java.lang.Long): Unit = {
+      val (toRemove, toKeep) = keys.asScala.partition(_ < cutoff)
+      if (toRemove.nonEmpty) {
+        toRemove.foreach(mapState.remove)
+        keysState.update(new java.util.ArrayList(toKeep.asJava))
+      }
+    }
+
+    def clear(): Unit = {
+      mapState.clear()
+      keysState.update(null)
+    }
+
+  }
 
 }
