+++ engine/flink/components/base-unbounded/src/main/scala/pl/touk/nussknacker/engine/flink/util/transformer/aggregate/EmitExtraWindowWhenNoDataTumblingAggregatorFunction.scala
@@ -1,7 +1,6 @@
 package pl.touk.nussknacker.engine.flink.util.transformer.aggregate
 
 import org.apache.flink.api.common.functions.{OpenContext, RuntimeContext}
-import org.apache.flink.api.common.state.ValueState
 import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.flink.streaming.api.TimerService
 import org.apache.flink.streaming.api.functions.KeyedProcessFunction
@@ -9,13 +8,8 @@ import org.apache.flink.util.Collector
 import pl.touk.nussknacker.engine.api.{Context => NkContext, NodeId, ValueWithContext}
 import pl.touk.nussknacker.engine.api.runtimecontext.{ContextIdGenerator, EngineRuntimeContext}
 import pl.touk.nussknacker.engine.api.typed.typing.TypingResult
-import pl.touk.nussknacker.engine.flink.api.state.StateHolder
-import pl.touk.nussknacker.engine.flink.util.keyed.{KeyEnricher, StringKeyedValue}
-import pl.touk.nussknacker.engine.flink.util.orderedmap.FlinkRangeMap
-import pl.touk.nussknacker.engine.flink.util.orderedmap.FlinkRangeMap._
+import pl.touk.nussknacker.engine.flink.util.keyed.KeyEnricher
 import pl.touk.nussknacker.engine.util.KeyedValue
-
-import scala.language.higherKinds
 
 /**
  * It behaves similar to Flink's TumblingWindow with one difference that we produce extra zero aggregate for each key when no data arrived.
@@ -23,7 +17,7 @@ import scala.language.higherKinds
  * it handles out of order elements. The other difference from AggregatorFunction is that we emit event only in timer and handle
  * state eviction on ours own.
  */
-class EmitExtraWindowWhenNoDataTumblingAggregatorFunction[MapT[K, V]](
+class EmitExtraWindowWhenNoDataTumblingAggregatorFunction(
     protected val aggregator: Aggregator,
     protected val timeWindowLengthMillis: Long,
     protected val timeWindowOffsetMillis: Long,
@@ -31,10 +25,8 @@ class EmitExtraWindowWhenNoDataTumblingAggregatorFunction[MapT[K, V]](
     protected val aggregateElementType: TypingResult,
     protected override val aggregateTypeInformation: TypeInformation[AnyRef],
     val convertToEngineRuntimeContext: RuntimeContext => EngineRuntimeContext
-)(implicit override val rangeMap: FlinkRangeMap[MapT])
-    extends KeyedProcessFunction[AnyRef, ValueWithContext[KeyedValue[AnyRef, AnyRef]], ValueWithContext[AnyRef]]
-    with StateHolder[MapT[Long, AnyRef]]
-    with AggregatorFunctionMixin[MapT] {
+) extends KeyedProcessFunction[AnyRef, ValueWithContext[KeyedValue[AnyRef, AnyRef]], ValueWithContext[AnyRef]]
+    with AggregatorFunctionMixin {
 
   type FlinkCtx =
     KeyedProcessFunction[AnyRef, ValueWithContext[KeyedValue[AnyRef, AnyRef]], ValueWithContext[AnyRef]]#Context
@@ -42,17 +34,19 @@ class EmitExtraWindowWhenNoDataTumblingAggregatorFunction[MapT[K, V]](
     KeyedProcessFunction[AnyRef, ValueWithContext[KeyedValue[AnyRef, AnyRef]], ValueWithContext[AnyRef]]#OnTimerContext
 
   @transient
-  protected var state: ValueState[MapT[Long, AnyRef]] = _
-
-  @transient
   private var contextIdGenerator: ContextIdGenerator = _
 
   override def open(openContext: OpenContext): Unit = {
-    state = getRuntimeContext.getState(stateDescriptor)
+    super.open(openContext)
+    initState()
     contextIdGenerator = convertToEngineRuntimeContext(getRuntimeContext).contextIdGenerator(nodeId.id)
   }
 
   override protected val minimalResolutionMs: Long = timeWindowLengthMillis
+
+  override protected def initEvictionTimeState(): Unit = {
+    // eviction is handled in onTimer — latestEvictionTimeForKey is not used
+  }
 
   override def processElement(
       value: ValueWithContext[KeyedValue[AnyRef, AnyRef]],
@@ -74,8 +68,7 @@ class EmitExtraWindowWhenNoDataTumblingAggregatorFunction[MapT[K, V]](
 
   override def onTimer(timestamp: Long, ctx: FlinkOnTimerCtx, out: Collector[ValueWithContext[AnyRef]]): Unit = {
     val previousTimestamp = timestamp - timeWindowLengthMillis
-    val currentStateValue = readStateOrInitial()
-    val finalVal          = computeFinalValue(currentStateValue, previousTimestamp)
+    val finalVal          = computeFinalValue(previousTimestamp, bucketsState.keys)
     out.collect(
       ValueWithContext(
         finalVal,
@@ -83,30 +76,17 @@ class EmitExtraWindowWhenNoDataTumblingAggregatorFunction[MapT[K, V]](
       )
     )
 
-    val previousTimestampStateAndRest = stateForTimestampToReadUntilEnd(currentStateValue, previousTimestamp)
-    if (previousTimestampStateAndRest.toScalaMapRO.isEmpty) {
-      evictStates()
-    } else {
+    val rangeStart  = previousTimestamp - timeWindowLengthMillis + 1
+    val hasMoreData = bucketsState.keys.stream().anyMatch(k => k >= rangeStart)
+    if (hasMoreData) {
       ctx.timerService().registerEventTimeTimer(timestamp + timeWindowLengthMillis)
+    } else {
+      bucketsState.clear()
     }
-  }
-
-  override protected def updateState(
-      stateValue: MapT[Long, AnyRef],
-      stateValidity: Long,
-      timeService: TimerService
-  ): Unit = {
-    state.update(stateValue)
   }
 
   override protected def doMoveEvictionTime(time: Long, timeService: TimerService): Unit = {
     // state eviction is controlled in onTimer
   }
-
-  protected def evictStates(): Unit = {
-    state.clear()
-  }
-
-  override protected def readState(): MapT[Long, AnyRef] = state.value()
 
 }
