+++ engine/flink/components/base-unbounded/src/main/scala/pl/touk/nussknacker/engine/flink/util/transformer/join/FullOuterJoinTransformer.scala
@@ -41,7 +41,6 @@ import pl.touk.nussknacker.engine.util.KeyedValue
 
 import java.time.Duration
 import java.util.concurrent.TimeUnit
-import scala.collection.immutable.SortedMap
 import scala.concurrent.duration.FiniteDuration
 import scala.jdk.CollectionConverters._
 
@@ -217,7 +216,7 @@ class FullOuterJoinTransformer(
   )(
       implicit nodeId: NodeId
   ): KeyedProcessFunction[String, ValueWithContext[StringKeyedValue[AnyRef]], ValueWithContext[AnyRef]] =
-    new FullOuterJoinAggregatorFunction[SortedMap](
+    new FullOuterJoinAggregatorFunction(
       aggregator,
       stateTimeout.toMillis,
       nodeId,
