+++ engine/flink/components/base-unbounded/src/main/scala/pl/touk/nussknacker/engine/flink/util/transformer/join/SingleSideJoinAggregatorFunction.scala
@@ -1,6 +1,6 @@
 package pl.touk.nussknacker.engine.flink.util.transformer.join
 
-import org.apache.flink.api.common.functions.RuntimeContext
+import org.apache.flink.api.common.functions.{OpenContext, RuntimeContext}
 import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.flink.streaming.api.functions.co.CoProcessFunction
 import org.apache.flink.util.Collector
@@ -8,37 +8,37 @@ import pl.touk.nussknacker.engine.api._
 import pl.touk.nussknacker.engine.api.NodeId
 import pl.touk.nussknacker.engine.api.runtimecontext.EngineRuntimeContext
 import pl.touk.nussknacker.engine.api.typed.typing.TypingResult
-import pl.touk.nussknacker.engine.flink.api.state.LatelyEvictableStateCoFunction
 import pl.touk.nussknacker.engine.flink.util.keyed.StringKeyedValue
-import pl.touk.nussknacker.engine.flink.util.orderedmap.FlinkRangeMap
 import pl.touk.nussknacker.engine.flink.util.transformer.aggregate.{Aggregator, AggregatorFunctionMixin}
+import pl.touk.nussknacker.engine.util.KeyedValue
 
-import scala.language.higherKinds
-
-class SingleSideJoinAggregatorFunction[MapT[K, V]](
+class SingleSideJoinAggregatorFunction(
     protected val aggregator: Aggregator,
     protected val timeWindowLengthMillis: Long,
     override val nodeId: NodeId,
     protected val aggregateElementType: TypingResult,
     override protected val aggregateTypeInformation: TypeInformation[AnyRef],
     val convertToEngineRuntimeContext: RuntimeContext => EngineRuntimeContext
-)(implicit override val rangeMap: FlinkRangeMap[MapT])
-    extends LatelyEvictableStateCoFunction[ValueWithContext[String], ValueWithContext[
+) extends CoProcessFunction[ValueWithContext[String], ValueWithContext[
       StringKeyedValue[AnyRef]
-    ], ValueWithContext[AnyRef], MapT[Long, AnyRef]]
-    with AggregatorFunctionMixin[MapT] {
+    ], ValueWithContext[AnyRef]]
+    with AggregatorFunctionMixin {
 
   type FlinkCtx = CoProcessFunction[ValueWithContext[String], ValueWithContext[
     StringKeyedValue[AnyRef]
   ], ValueWithContext[AnyRef]]#Context
+
+  override def open(openContext: OpenContext): Unit = {
+    super.open(openContext)
+    initState()
+  }
 
   override def processElement1(
       in1: ValueWithContext[String],
       ctx: FlinkCtx,
       out: Collector[ValueWithContext[AnyRef]]
   ): Unit = {
-    val current: MapT[Long, aggregator.Aggregate] = readStateOrInitial()
-    val finalVal                                  = computeFinalValue(current, ctx.timestamp())
+    val finalVal = computeFinalValue(ctx.timestamp(), bucketsState.keys)
     out.collect(ValueWithContext(finalVal, in1.context))
   }
 
@@ -47,7 +47,22 @@ class SingleSideJoinAggregatorFunction[MapT[K, V]](
       ctx: FlinkCtx,
       out: Collector[ValueWithContext[AnyRef]]
   ): Unit = {
-    addElementToState(in2, ctx.timestamp(), ctx.timerService(), out)
+    addElementToState(
+      in2.asInstanceOf[ValueWithContext[KeyedValue[AnyRef, AnyRef]]],
+      ctx.timestamp(),
+      ctx.timerService(),
+      out
+    )
+  }
+
+  override def onTimer(
+      timestamp: Long,
+      ctx: CoProcessFunction[ValueWithContext[String], ValueWithContext[
+        StringKeyedValue[AnyRef]
+      ], ValueWithContext[AnyRef]]#OnTimerContext,
+      out: Collector[ValueWithContext[AnyRef]]
+  ): Unit = {
+    handleOnTimer(timestamp, ctx.timerService)
   }
 
 }
