+++ engine/flink/components/base-unbounded/src/main/scala/pl/touk/nussknacker/engine/flink/util/transformer/join/SingleSideJoinTransformer.scala
@@ -31,7 +31,6 @@ import pl.touk.nussknacker.engine.util.KeyedValue
 
 import java.time.Duration
 import java.util.concurrent.TimeUnit
-import scala.collection.immutable.SortedMap
 import scala.concurrent.duration.FiniteDuration
 
 class SingleSideJoinTransformer(
@@ -205,7 +204,7 @@ class SingleSideJoinTransformer(
   )(
       implicit nodeId: NodeId
   ): CoProcessFunction[ValueWithContext[String], ValueWithContext[StringKeyedValue[AnyRef]], ValueWithContext[AnyRef]] =
-    new SingleSideJoinAggregatorFunction[SortedMap](
+    new SingleSideJoinAggregatorFunction(
       aggregator,
       stateTimeout.toMillis,
       nodeId,
